+++ src/main/java/uk/gov/hmcts/reform/slc/services/servicebus/MessageProcessor.java
@@ -7,6 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.slc.services.SendLetterJob;
+import uk.gov.hmcts.reform.slc.services.servicebus.exceptions.ConnectionException;
 
 import java.util.function.Function;
 import java.util.function.Supplier;
