+++ src/main/java/uk/gov/hmcts/reform/slc/services/servicebus/exceptions/ConnectionException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.slc.services.servicebus;
+package uk.gov.hmcts.reform.slc.services.servicebus.exceptions;
 
 public class ConnectionException extends RuntimeException {
     public ConnectionException(String message, Throwable cause) {
