+++ src/main/java/uk/gov/hmcts/reform/slc/services/steps/maptoletter/LetterMapper.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.microsoft.azure.servicebus.IMessage;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.slc.model.Letter;
+import uk.gov.hmcts.reform.slc.services.steps.maptoletter.exceptions.InvalidMessageException;
 
 import java.io.IOException;
